projects/analytics/airregi-analytics/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
pytz>=2023.3
pydantic>=2.5.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.12.0
flake8>=7.0.0
